Build allows Open-Quark and CAL-Eclipse-Plug-in to be in separate folders now.
Previously, the Ant build would only succeed if the CAL-Eclipse-Plug-in code was in the Open-Quark folder.
This was problematic for file management and source control since some of the files and folders overlapped.

Now, a separate property (cal.eclipse.plugin.src.folder) is used to specify the CAL plug-in source location.
By default, this is set to the folder "CAL-Eclipse-Plug-in/src" where the "CAL-Eclipse-Plug-in" folder is at the same level as the "Open-Quark" folder.
This can be overridded using the Ant property "cal.eclipse.plugin.src.folder" if necessary.